requirements.txt
# external requirements
#click
#Sphinx
#coverage
pytest-cov
biopython==1.79
ete3==3.1.2
google-cloud-bigquery
google-cloud-storage
numpy==1.21.5
pandas==1.3.5
requests
tqdm
black[jupyter]
